Update README.md with more information on the App and restructured features section
This commit adds more detailed information on how the Grocery App was built, including the addition of VueJS and Vuetify. It also reorganizes the 'Features' into three sections for users, admins, and managers. This provides a clearer overview of the app's functionality for different roles. Furthermore, the steps for project setup have been updated with a new 'Clone the repository' section and revised compile instructions. This should give potential contributors a more comprehensive guide on how to get started with the project.