Remove AboutView.vue from the project
This commit deletes the AboutView.vue from the source code because the 'About' section is no longer necessary in the project. The login form that was originally part of the `AboutView` is moved to a different section of the website. This change simplifies the site navigation and improves the user interface.