Refactor App.vue to use Navbar component
Changed the structure of App.vue to utilize the newly created Navbar component. This change cleans up the template by removing the navigation links earlier implemented in App.vue itself. Now, all the navigation related markup is encapsulated in the Navbar component. This makes the code more maintainable and modular.