Add registration component and view
Created RegisterView and RegistrationForm components.
The RegisterView serves as a container for the RegistrationForm.
RegistrationForm contains input fields for user registration, such as username, password, and email.
Implemented vee-validate for input validation.
This allows users to sign up in order to use the application, and prevents the submission of invalid input data.