Improve User Authentication and Logout Functionality
The commit carries out several changes to improve user login/logout process.

Firstly, a new method "logout" was added to "Navbar.vue" allowing user to safely logout by clearing user state and access token. The link was changed to a div which on click triggers the logout method.

Secondly, in "App.vue", changes were made to check on every mount, whether the user is logged in by invoking the "checkLogin" method that verifies the access token.

Thirdly, improvements to "baseStore.js" include several new methods like "checkLogin" and "logout" to handle the user login state and access token from cookies.

Additionally, the cookie setting method was modified to include flags "SameSite=None; Secure;". This change increases the security of the cookie.

Last but not least, the method "fetchAccessToken" was refactored to be an async function, making the fetching of an access token a more efficient process. It now uses the fetch API instead of axios which simplifies the code and makes it easier to handle catch errors.

All these measures make the user authentication process more efficient and secure and enhance the user experience.