Update dependencies and resolve vulnerabilities
Update multiple package versions in package-lock.json to mitigate vulnerabilities and ensure compatibility with the latest features. Specifically, the version of vue is updated from 3.3.4 to 3.3.8, eslint is updated from 8.52.0 to 8.53.0, and several other dependencies are updated as well.

This will enhance the security and performance of the application. Changes also include updates to checksums and resolved URLs.