Update signup button functionality in LoginForm.vue
Adjusted the signup button's functionality in LoginForm.vue. Instead of redirecting to a separate register route on click, it now emits a custom event 'signup-clicked'. This change was made to potentially allow parent components to dynamically toggle the visibility of the RegistrationForm, improving user experience and UI flow. Additionally, the signup button is not displayed if the user type is 'admin'.