Add an option to open as user and revise visibility of admin and manager buttons
Implemented a new feature that enables opening the app as a 'user'. Added related functionality that signs out the current user and re-routes to the homepage. Also revised the visibility of 'admin' and 'manager' buttons such that they only appear when the current user type is not 'admin' or 'manager' respectively. This provides a more intuitive interface to the users.