Update access control and CSS variables in AdminView.vue
Admin dashboard access security was updated to permit access to the admin dashboard only for users of type 'admin'. This change reduces the possibility of unauthorized access. Also, CSS properties for #adminLogin and #adminDashboard were refactored to use global variables for width and height. This enhances code consistency and affords easier updates in the future.