Adjust Navbar.vue CSS properties and padding
The padding and width properties under .navbar have been updated to improve the overall look and responsiveness of the navbar. The padding was removed and width was refractored to rely on a global variable (--app-width) instead of a fixed value. Additionally, height property was added to make the navbar height consistent across the application.