Update admin dashboard UI and improve navigation
The admin dashboard UI was updated to enhance its readability. Buttons for approving or rejecting manager requests were re-enabled and their styles were updated for better visibility. Table element styles were also modified to centralize the alignments. New 'CategoryForm' component was added to be utilized for category management tasks.
To improve user experience, navigation was also tweaked in the registration form so that the user is redirected to the homepage upon successful registration after a delay of 5 seconds, rather than the immediate redirect to login page. Additionally, commented out background colors in the AdminDashboard component.