Implement Category Request Form in ManagerDashboard
Introduced a new component 'CategoryRequestForm' in the ManagerDashboard. This form allows managers to make category requests. Form validation has been added to ensure category name and description meet certain criteria before submission. A name for a category is required to be between 4 and 20 characters while a description should be between 10 and 20 characters. If the request is successful, a message notifies the manager that the request was created and needs admin approval. Errors in the request process are also handled and displayed to the manager. This feature was added to give managers more autonomy on site content management.