Update package.json and package-lock.json
Upgraded dependencies to their latest versions in package and package-lock json files.

This commit upgrades "@bootstrap-icons" , "vee-validate", and "@rushstack/eslint-patch" to their latest respective versions. This update was necessary to ensure that we are using the most up-to-date and secure versions of our dependencies. This can potentially improve app performance, fix bugs and prevent potential security vulnerabilities.