Update ManagerDashboard component for improved UI and functionality
Added `productCard` and `categoryList` components to the `ManagerDashboard` component and updated its UI for betterment. Removed the older way of displaying categories and products, and adopted a new approach where each category and product is represented by its own card, improving visibility and UX. Load of the categories is now handled by the `categoryList` component. Additionally, adjusted proportion of the left and right side in the dashboard and added styling to the newly added cards. ManagerDashboard functionality and appearance is thus enhanced.