Add Notification component with alert system\n
Implemented a new 'Notification' component which is integrated into the 'App' component as well as the 'baseStore'. This component accepts props to generate notifications alerts based on type attributes like 'success', 'danger', and 'info'. This aims to enhance user interaction and user experience by providing appropriate, timely feedback to the user. A hide method is also included to allow users to dismiss the notification.