Remove CategoryForm.vue and incorporate its functionality in CategoryTable.vue
Deleted the CategoryForm.vue component and restructured the CategoryTable.vue by embedding the form that was earlier present in CategoryForm.vue. This change was made to make the 'create category' flow more intuitive and part of a single component, thus enhancing user experience. The layout of categories is also updated in terms of UI/UX. All methods for 'category' previously in CategoryForm.vue are now in CategoryTable.vue, ensuring functionality stays intact. Update AdminDashboard.vue to remove CategoryForm.vue reference.