Add 'type' parameter to 'approveCategoryRequest' function
Added an additional 'type' parameter to 'approveCategoryRequest' function in 'adminStore.js' to handle both 'add' and 'delete' category actions. This change allows the function to distinguish between a request to add or delete a category, thus enabling more granulated user actions and smoother UI/UX. To ensure a consistent flow, the 'fetchCategoryRequests' method is now called regardless of the request type. Additionally, if the 'delete' request is successful/unsuccessful, appropriate notifications are shown.