Updated CRUD operations in CategoryTable.vue component
Introduced changes in the CategoryTable.vue component to improve user interactions with category data. Previously, the category data object was not well-structured which was causing issues in POST requests. Accordingly, the data object was split into three separate objects: 'createCat' for creating a new category, 'editCat' for editing existing category, and 'toDelete' for deleting a category. The UI changes were done to enable edit operations directly in the table cells and handle data operations in a more intuitive way. On success or failure of these operations, appropriate notifications are shown to the user.