Update login check logic and improve UX
Refactored the login check function in baseStore.js to return both the login status and user type information. This change eliminates redundant console logs and enhances the efficiency of the function by providing the necessary data for further operations.

Moreover, this modification impacts the AdminView.vue execution as the checkLogin function response is now an array. Therefore, updated the condition in the beforeMount lifecycle hook in AdminView.vue to handle the new response structure.

This change results in not only more streamlined and efficient code but also a more user-friendly admin view as stakeholder type is handled more robustly.