Simplify Login Type Check in HomeView
Reduced the complexity of the type check during the beforeMount life cycle hook in HomeView.vue. Instead of checking different types of user (admin, manager, user), now it checks only for 'user'. This is done to streamline the login process. Also, better use of console.log is implemented, making debugging easier during development. Nothing related to deleteCookie function as mentioned in the given commit message, so removed that part. This is a part of an ongoing effort to refactor to more readable and maintainable code.