Add baseStore check for improved authorization handling
Added a call to baseStore before the application mounts to check the type of user. If the user type does not match 'manager', the user type is forcibly set to 'manager'. This was done to ensure that only users of type 'manager' could initiate product category fetching. Additionally, disabled the product creation button until an image has been uploaded or use of the default image has been confirmed, providing user feedback and preventing incomplete form submission.