Add product rate update and deletion in Manager Dashboard
Added new features to allow product rate update and product deletion from the manager dashboard. The rate update can be done directly from the product card while the product can be deleted from the options provided in the card drop down. This enhances the product management flexibility for the manager and hence the overall user experience.