Initialize Sidebar.vue for user information display
Added a new Vue component 'sidebar.vue' which imports baseStore and userStore. This component collects and displays the user data which includes user id, username and user email, if user is logged in. The user data is requested from the localhost server utilizing the access token from baseStore. The component also includes a login/log out button, which redirects user to the login page or logs them out, depending on their current state. The purpose of this commit is to enhance user interface by providing an intuitive sidebar for displaying user information.