Update Navbar.vue to improve user navigation
Navbar.vue has been updated to improve how the user navigates through the application. A new computed property 'home' has been added to handle redirection of user based on their role after logging in. If the user role type is 'admin', they are redirected to '/admin', if 'manager' to '/manager', otherwise it defaults to home '/'. The navbar-brand href is bound to this new property ensuring dynamic redirection. Furthermore, the navbar structure has been simplified and the logout button is now rendered only if the user is logged in. These changes aim to enhance the user experience by offering more intuitive navigation and role specific routing.