Redirect unauthorized users in AddProductView.vue
Modified error handling logic in AddProductView.vue. Unauthenticated users trying to access the manager's view will now be logged out and redirected to the home page. This change was necessary to ensure secure access and prevent unauthorized access to manager functionalities.

Additionally, renamed and relocated ManagerDashboard.vue to the 'manager dashboard components' directory as part of a larger file restructuring effort for better organization and clarity of project components.