Add getter for categoryRequests in adminStore
A new getter method has been added to the 'adminStore.js' to allow access to the 'categoryRequests' array. This array was previously fetched directly from the adminStore. Now, it is fetched through the getter method 'getCategoryRequests'. This change will provide better encapsulation and interfaces to interact with the store data, making the codebase more maintainable and scalable.