Integrate authentication checks and user experience improvements to ProductCard.vue
Imported 'baseStore' and 'userStore', set up baseStore and userStore in setup() and used them in a new toCart() method. This method is used to control users' action of adding products to the cart, ensuring only logged-in users can do so. Additionally, the display of product price and current stock information has been improved. The quantity selection feature has also been enhanced; users can control the quantity of the product they want to buy in a more interactive and intuitive way. All of these changes aim to enhance the overall user experience.