Improve sidebar display and UX for logged-in users
Added detailed styles for brand logo, cart and buttons on the sidebar for better user experience. Incorporated authentication checks - display of user details and specific buttons are now controlled by the logged-in status. Consequently, the user interface became more user-friendly and clearer, particularly for logged-in users. Styling has been applied to the components for a more appealing display.