Improve product card with out of stock feature
Added a computed property to determine if a product is out of stock, and implemented UI changes to denote the product's status. Updated the look and functionality of the product card for both managers and users: for managers, better visual elements were added for the product's state of availability, and for users, fetch functions for unconfirmed and confirmed orders were added.