Fix order confirmation status in userStore
Fixed a bug that prevented orders from being correctly marked as confirmed in the user store. Previously, orders were mistakenly added to the "unconfirmed" array, but they are now correctly added to the "confirmed" array when the response is ok. This resolves an issue which caused confirmed orders to be inaccurately represented in the application.